[SPARK-55792][PS] Fix CI: ruff format, lag window frame, groupby diff dtype
- Apply ruff format corrections to frame.py (dict comprehension layout, slice spacing)

- Remove rowsBetween from lag window in Series._diff for Spark Connect compatibility

- Update test_groupby_diff expectations to float dtype (remove .astype(int) cast)

+++ python/pyspark/pandas/tests/diff_frames_ops/test_groupby_diff.py
@@ -57,7 +57,7 @@ class GroupByDiffMixin:
             pdf.groupby(pkey)[["a"]].diff().sort_index(),
         )
 
-        self.assert_eq(psdf.groupby(kkey).diff().sum(), pdf.groupby(pkey).diff().sum().astype(int))
+        self.assert_eq(psdf.groupby(kkey).diff().sum(), pdf.groupby(pkey).diff().sum())
         self.assert_eq(psdf.groupby(kkey)["a"].diff().sum(), pdf.groupby(pkey)["a"].diff().sum())
